pyproject.toml
[project]
name = "projeto-db4"
version = "0.1.0"
description = ""
authors = [
    {name = "João Dorea",email = "84815558+joaopdorea@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas (>=2.3.1,<3.0.0)",
    "sqlalchemy (>=2.0.42,<3.0.0)",
    "yfinance (>=0.2.65,<0.3.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "dbt-postgres (>=1.9.0,<2.0.0)",
    "streamlit (>=1.48.0,<2.0.0)",
    "streamlit-authenticator (>=0.4.2,<0.5.0)"
]

[tool.poetry]
packages = [{include = "projeto_db4", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
